finished player input buffer
player input is now such that: no two keys can be registered at the same time; any secondary input overrides old input but old input is restored when secondary input is released. say, for instance, the player is pressing the d key. the character will move to the right. now, without letting go of the d key, the player starts pressing the w key. the character will move up, until the player lets go of the w key, when the player will start moving to the right again, because the d key is still being pressed.